Add `CacheableMock` and unit tests for new CacheableModel manager
This override is required because the get use the order_by() by default
and sends an unused query in case of cached model is found.

I added tests on queries count. When an object is saved and recalled in
a second time, the filter query must return the cached object itself
instead of super.